+++ src/R.NET/NumericVector.cs
@@ -97,7 +97,7 @@ namespace RDotNet
         /// <returns>The element at the specified index.</returns>
         protected override double GetValueAltRep(int index)
         {
-            return GetFunction<REAL_ELT>()(DangerousGetHandle(), index);
+            return GetFunction<REAL_ELT>()(DangerousGetHandle(), (IntPtr)index);
         }
 
         /// <summary>
@@ -122,7 +122,7 @@ namespace RDotNet
         /// <param name="value">The value to set</param>
         protected override void SetValueAltRep(int index, double value)
         {
-            GetFunction<SET_REAL_ELT>()(DangerousGetHandle(), index, value);
+            GetFunction<SET_REAL_ELT>()(DangerousGetHandle(), (IntPtr)index, value);
         }
 
         /// <summary>
@@ -145,7 +145,7 @@ namespace RDotNet
         {
             // by inference from `static SEXP compact_intseq_Duplicate(SEXP x, Rboolean deep)`  in altrep.c
             var res = new double[Length];
-            GetFunction<REAL_GET_REGION>()(DangerousGetHandle(), 0, Length, res);
+            GetFunction<REAL_GET_REGION>()(DangerousGetHandle(), (IntPtr)0, (IntPtr)Length, res);
             return res;
         }
